+++ Net_Core/Web.Angular/projects/gw-lib/core/file-manager/src/interfaces/upload-status.model.ts
@@ -1,5 +1,6 @@
 export interface IUploadStatus {
     id: string;
+    fileId: string;
     fileName: string;
     message: string;
     completed: boolean;
@@ -11,6 +12,7 @@ export interface IUploadStatus {
 export class UploadStatus implements IUploadStatus {
 	constructor(
       public id: string,
+      public fileId: string,
       public fileName: string,
       public message: string,
       public completed: boolean,